+++ frontend/src/components/form/SubmitButton.js
@@ -1,6 +1,6 @@
 import useFormContext from "../../hooks/useFormContext";
 import PropTypes from "prop-types";
-import Button from "../../components/@common/Button/Button";
+import Button from "../@common/Button/Button";
 
 const SubmitButton = ({ children }) => {
   const { isValid, isEmpty } = useFormContext();
